+++ HISTORY.rst
@@ -2,6 +2,21 @@
 History
 =======
 
+0.3.0 (2017-03-22)
+------------------
+
+* README getting started section.
+* Tests to state machine without model.
+
+
+0.2.0 (2017-03-22)
+------------------
+
+* ``State`` can hold a value that will be assigned to the model as the state value.
+* Travis-CI integration.
+* RTD integration.
+
+
 0.1.0 (2017-03-21)
 ------------------